refactor enum

+++ Metrics.Core.MVC/MastodonService.cs
@@ -1,8 +1,8 @@
 ﻿using Mastodon.Api;
 using Mastodon.Model;
-using Metrics.Core.Enum;
 using Metrics.Core.Service;
 using Metrics.Model;
+using Metrics.Model.Enum;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
@@ -28,14 +28,14 @@ namespace Metrics.Core.MVC
         {
             await Setup();
             log.LogInformation("{Count} {username}", user.FollowersCount, username);
-            return await Chart.SaveData(user.FollowersCount, (int)MetricType.MastodonFollowers, username);
+            return await Chart.SaveData(user.FollowersCount, MetricType.MastodonFollowers, username);
         }
 
         public async Task<IActionResult> GetMastodonFollowing(ILogger log, string username)
         {
             await Setup();
             log.LogInformation("{Count} {username}", user.FollowingCount, username);
-            return await Chart.SaveData(user.FollowingCount, (int)MetricType.MastodonFollowing, username);
+            return await Chart.SaveData(user.FollowingCount, MetricType.MastodonFollowing, username);
         }
 
         public async Task GetFollowFriday(ILogger log)
@@ -135,7 +135,7 @@ namespace Metrics.Core.MVC
         {
             await Setup();
             log.LogInformation("{Count} {username}", user.StatusesCount, username);
-            return await Chart.SaveData(user.StatusesCount, (int)MetricType.NumberOfToots, username);
+            return await Chart.SaveData(user.StatusesCount, MetricType.NumberOfToots, username);
         }
 
         private async Task Setup()
